Adapt to feature/reworkRequireJs in assetgraph.
buildProduction: Stop using removeOrPolyfillSystemJs now that it's part of the bundleSystemJs transform.

Remove some require.js test cases that will be unsupported as they don't use data-main or expect cross-page bundling.
Avoid root-relative paths in the regression test for issue #58. I cannot make it work as-is with the require.js optimizer [ci skip].
Adapt the 'simple' buildProduction test to not use the tpl plugin [ci skip].